Spostati alcuni metodi della gestione dei voti e il metodo generale NextKey
