+++ BusinessLayer/BL_GradesManagement.cs
@@ -1,0 +1,13 @@
+﻿using SchoolGrades.DbClasses;
+using System;
+using System.Collections.Generic;
+using System.Data.Common;
+using System.Text;
+
+namespace SchoolGrades
+{
+    internal partial class BusinessLayer
+    {
+        
+    }
+}
