+++ DbClasses/DbAndBusiness.cs
@@ -17,6 +17,7 @@ namespace SchoolGrades.DbClasses
     public class DbAndBusiness
     {
         DataLayer dl;
+        BusinessLayer bl;
         private string dbName;
 
         public string DatabaseName { get => dbName; }
@@ -24,8 +25,8 @@ namespace SchoolGrades.DbClasses
         #region constructors
         public DbAndBusiness(string PathAndFile)
         {
-            dl = new DataLayer(PathAndFile); 
-
+            dl = new DataLayer(PathAndFile);
+            bl = new BusinessLayer(PathAndFile);
             if (!System.IO.File.Exists(PathAndFile))
             {
                 string err = @"[" + PathAndFile + " not in the current nor in the dev directory]";
@@ -35,41 +36,7 @@ namespace SchoolGrades.DbClasses
             dbName = PathAndFile;
         }
         #endregion
-        internal DataTable GetGradesOfStudent(Student Student, string SchoolYear,
-            string IdGradeType, string IdSchoolSubject,
-            DateTime DateFrom, DateTime DateTo)
-        {
-            DataTable t;
-            using (DbConnection conn = dl.Connect())
-            {
-                string query = "SELECT Grades.idGrade,datetime(Grades.timeStamp), Students.idStudent," +
-                "lastName,firstName," +
-                " Grades.value AS 'grade', Grades.weight," +
-                " Grades.idGradeParent" +
-                " FROM Grades" +
-                " JOIN Students" +
-                " ON Students.idStudent=Grades.idStudent" +
-                " JOIN Classes_Students" +
-                " ON Classes_Students.idStudent=Students.idStudent" +
-                " WHERE Students.idStudent =" + Student.IdStudent +
-                " AND Grades.idSchoolYear='" + SchoolYear + "'" +
-                " AND Grades.idGradeType = '" + IdGradeType + "'" +
-                " AND Grades.idSchoolSubject = '" + IdSchoolSubject + "'" +
-                " AND Grades.Value > 0" +
-                " AND Grades.Timestamp BETWEEN " + SqlVal.SqlDate(DateFrom) + " AND " + SqlVal.SqlDate(DateTo) +
-                " ORDER BY lastName, firstName, Students.idStudent, Grades.timestamp Desc;";
-
-                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                DataSet DSet = new DataSet("ClosedMicroGrades");
-
-                DAdapt.Fill(DSet);
-                t = DSet.Tables[0];
-
-                DAdapt.Dispose();
-                DSet.Dispose();
-            }
-            return t;
-        }
+        
         internal void FixQuestionInGrade(int? IdGrade)
         {
             using (DbConnection conn = dl.Connect())
@@ -99,12 +66,6 @@ namespace SchoolGrades.DbClasses
             }
         }
 
-        internal object GetGradesWeightsOfStudentOnOpenGrades(Student currentStudent, string stringKey1, string stringKey2, DateTime value1, DateTime value2)
-        {
-            throw new NotImplementedException();
-
-        }
-
         internal void UpdatePathStartLinkOfClass(Class currentClass, string text)
         {
             // !!!! currently not used, because pathStartLink field does not exist yet in the database !!!!
@@ -120,52 +81,6 @@ namespace SchoolGrades.DbClasses
                 cmd.ExecuteNonQuery();
                 cmd.Dispose();
             }
-        }
-
-        internal DataTable GetGradesWeightsOfClassOnOpenGrades(Class Class,
-            string IdGradeType, string IdSchoolSubject, DateTime DateFrom, DateTime DateTo)
-        {
-            DataTable t;
-            using (DbConnection conn = dl.Connect())
-            {
-                // find the macro grade type of the micro grade
-                // TODO take it from a Grade passed as parameter 
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT idGradeTypeParent " +
-                    "FROM GradeTypes " +
-                    "WHERE idGradeType='" + IdGradeType + "'; ";
-                string idGradeTypeParent = (string)cmd.ExecuteScalar();
-
-                string query = "SELECT Grades.idGrade,Students.idStudent,lastName,firstName" +
-                ",SUM(Grades.weight)/100 AS 'GradesFraction', 1 - SUM(Grades.weight)/100 AS LeftToCloseAssesments" +
-                ",COUNT() AS 'GradesCount'" +
-                " FROM Grades, Grades AS Parents " +
-                " JOIN Classes_Students ON Students.idStudent=Grades.idStudent" +
-                " JOIN Students ON Classes_Students.idStudent=Students.idStudent" +
-                " WHERE Classes_Students.idClass =" + Class.IdClass +
-                " AND Grades.idSchoolYear='" + Class.SchoolYear + "'" +
-                " AND (Grades.idGradeType='" + IdGradeType + "'" +
-                " OR Grades.idGradeType IS NULL)" +
-                " AND Grades.idSchoolSubject='" + IdSchoolSubject + "'" +
-                " AND Grades.value IS NOT NULL AND Grades.value <> 0" +
-                " AND Grades.Timestamp BETWEEN " + SqlVal.SqlDate(DateFrom) + " AND " + SqlVal.SqlDate(DateTo) +
-                " AND Parents.idGradeType = '" + idGradeTypeParent + "'" +
-                " AND Grades.idGradeParent = Parents.idGrade" +
-                " AND (Parents.Value is null or Parents.Value = 0)" +
-                " AND NOT Students.disabled" +
-                " GROUP BY Students.idStudent" +
-                " ORDER BY GradesFraction ASC, lastName, firstName, Students.idStudent;";
-
-                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                DataSet DSet = new DataSet("ClosedMicroGrades");
-
-                DAdapt.Fill(DSet);
-                t = DSet.Tables[0];
-
-                DAdapt.Dispose();
-                DSet.Dispose();
-            }
-            return t;
         }
 
         internal void SaveTableOnCvs(DataTable Table, string FileName)
@@ -268,7 +183,7 @@ namespace SchoolGrades.DbClasses
             string IdSchool)
         {
             // trova una chiave da assegnare alla nuova classe
-            int idClass = NextKey("Classes", "idClass");
+            int idClass = dl.NextKey("Classes", "idClass");
             using (DbConnection conn = dl.Connect())
             {
                 DbCommand cmd = conn.CreateCommand();
@@ -281,7 +196,7 @@ namespace SchoolGrades.DbClasses
                     "');";
                 cmd.ExecuteNonQuery();
 
-                int nextId = NextKey("Classes_StartLinks", "idStartLink");
+                int nextId = dl.NextKey("Classes_StartLinks", "idStartLink");
                 cmd = conn.CreateCommand();
                 // create a link in StartLinks' link table
                 cmd.CommandText = "INSERT INTO Classes_StartLinks " +
@@ -299,36 +214,7 @@ namespace SchoolGrades.DbClasses
                 Directory.CreateDirectory(Commons.PathImages + "\\" + SchoolYear + ClassAbbreviation);
             }
             return idClass;
-        }
-
-        internal DataTable GetUnfixedGrades(Student Student, string IdSchoolSubject,
-            double Threshold)
-        {
-            DataTable t;
-            using (DbConnection conn = dl.Connect())
-            {
-                DataAdapter dAdapt;
-                DataSet dSet = new DataSet();
-                string query = "SELECT Grades.IdGrade,Grades.idStudent,Grades.value,Grades.timestamp,Grades.isFixed," +
-                    "Grades.idGradeType,Grades.idQuestion,Questions.text,Questions.*,Grades.*" +
-                    " FROM Grades" +
-                    " JOIN Questions ON Grades.idQuestion=Questions.idQuestion" +
-                    " WHERE idStudent=" + Student.IdStudent +
-                    " AND Grades.value<" + Threshold.ToString() +
-                    " AND (Grades.isFixed=0 OR Grades.isFixed is NULL)";
-                if (IdSchoolSubject != "")
-                    query += " AND Grades.idSchoolSubject='" + IdSchoolSubject + "'";
-                query += ";";
-                dAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                dSet = new DataSet("GetUnfixedGradesInTheYear");
-                dAdapt.Fill(dSet);
-                t = dSet.Tables[0];
-
-                dSet.Dispose();
-                dAdapt.Dispose();
-            }
-            return t;
-        }
+        }        
 
         private Question GetQuestionFromRow(DbDataReader Row)
         {
@@ -408,7 +294,7 @@ namespace SchoolGrades.DbClasses
         internal int CreateStudentFromStringMatrix(string[,] StudentData, int? StudentRow)
         {
             // trova una chiave da assegnare al nuovo studente
-            int codiceStudente = NextKey("Students", "idStudent");
+            int codiceStudente = dl.NextKey("Students", "idStudent");
             using (DbConnection conn = dl.Connect())
             {
                 DbCommand cmd = conn.CreateCommand();
@@ -434,7 +320,7 @@ namespace SchoolGrades.DbClasses
         internal int CreateStudent(Student Student)
         {
             // trova una chiave da assegnare al nuovo studente
-            int idStudent = NextKey("Students", "idStudent");
+            int idStudent = dl.NextKey("Students", "idStudent");
             Student.IdStudent = idStudent;
             using (DbConnection conn = dl.Connect())
             {
@@ -708,7 +594,7 @@ namespace SchoolGrades.DbClasses
             File.Move(PathAndFileName + "TEMP", newFileName);
 
             // trova la chiave per la prossima foto 
-            int codiceFoto = NextKey("StudentsPhotos", "idStudentsPhoto");
+            int codiceFoto = dl.NextKey("StudentsPhotos", "idStudentsPhoto");
             using (DbConnection conn = dl.Connect())
             {
                 DbCommand cmd = conn.CreateCommand();
@@ -747,7 +633,7 @@ namespace SchoolGrades.DbClasses
             // creation of a new class in the Classes table
 
             // finds a key for the new class
-            int idClass = NextKey("Classes", "idClass");
+            int idClass = dl.NextKey("Classes", "idClass");
             using (DbConnection conn = dl.Connect())
             {
                 DbCommand cmd = conn.CreateCommand();
@@ -760,9 +646,9 @@ namespace SchoolGrades.DbClasses
                 cmd.ExecuteNonQuery();
 
                 // find the key for next student
-                int idNextStudent = NextKey("Students", "idStudent");
+                int idNextStudent = dl.NextKey("Students", "idStudent");
                 // find the key for next picture 
-                int idNextPhoto = NextKey("StudentsPhotos", "idStudentsPhoto");
+                int idNextPhoto = dl.NextKey("StudentsPhotos", "idStudentsPhoto");
                 // add the student to the students' table 
                 // start from the second row of the file, first row is descriptions 
                 for (int riga = 1; riga < StudentsData.GetLength(0); riga++)
@@ -816,28 +702,6 @@ namespace SchoolGrades.DbClasses
                 cmd.Dispose();
             }
             return idClass;
-        }
-
-
-        private int NextKey(string Table, string Id)
-        {
-            int nextId;
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT MAX(" + Id + ") FROM " + Table + ";";
-                var firstColumn = cmd.ExecuteScalar();
-                if (firstColumn != DBNull.Value)
-                {
-                    nextId = int.Parse(firstColumn.ToString()) + 1;
-                }
-                else
-                {
-                    nextId = 1;
-                }
-                cmd.Dispose();
-            }
-            return nextId;
         }
 
         internal DataTable GetClassTable(int? idClass)
@@ -1108,7 +972,7 @@ namespace SchoolGrades.DbClasses
                     }
                     else
                     {
-                        IdStartLink = NextKey("Classes_StartLinks", "IdStartLink");
+                        IdStartLink = dl.NextKey("Classes_StartLinks", "IdStartLink");
                         cmd.CommandText = "INSERT INTO Classes_StartLinks" +
                             " (idStartLink,idClass,startLink,desc)" +
                             " VALUES " +
@@ -1199,45 +1063,6 @@ namespace SchoolGrades.DbClasses
             return listOfLinks;
         }
 
-        internal Grade GetGrade(int? IdGrade)
-        {
-            Grade g = null;
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT  * FROM Grades" +
-                    " WHERE Grades.idGrade=" + IdGrade.ToString() +
-                    ";";
-                dRead = cmd.ExecuteReader();
-                dRead.Read(); 
-                g = GetGradeFromRow(dRead);
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return g;
-        }
-
-        private Grade GetGradeFromRow(DbDataReader Row)
-        {
-            Grade g = new Grade();
-            g.IdGrade = (int)Row["idGrade"];
-            g.IdGradeParent = SafeDb.SafeInt(Row["idGradeParent"]);
-            g.IdStudent = SafeDb.SafeInt(Row["idStudent"]);
-            g.IdGradeType = SafeDb.SafeString(Row["IdGradeType"]);
-            g.IdSchoolSubject = SafeDb.SafeString(Row["IdSchoolSubject"]);
-            //g.IdGradeTypeParent = SafeDb.SafeString(Row["idGradeTypeParent"]);
-            g.IdQuestion = SafeDb.SafeInt(Row["idQuestion"]);
-            g.Timestamp = (DateTime)Row["timestamp"];
-            g.Value = SafeDb.SafeDouble(Row["value"]);
-            g.Weight = SafeDb.SafeDouble(Row["weight"]);
-            g.CncFactor = SafeDb.SafeDouble(Row["cncFactor"]);
-            g.IdSchoolYear = SafeDb.SafeString(Row["idSchoolYear"]);
-            //g.DummyInt = (int)Row["dummyInt"]; 
-            return g; 
-        }
-
         internal void GetGradeAndStudent(Grade Grade, Student Student)
         {
             using (DbConnection conn = dl.Connect())
@@ -1252,7 +1077,7 @@ namespace SchoolGrades.DbClasses
                 dRead = cmd.ExecuteReader();
                 while (dRead.Read())
                 {
-                    Grade = GetGradeFromRow(dRead);
+                    Grade = dl.GetGradeFromRow(dRead);
                     Student = GetStudentFromRow(dRead);
                     break; // just the first! 
                 }
@@ -1327,7 +1152,7 @@ namespace SchoolGrades.DbClasses
                 dRead = cmd.ExecuteReader();
                 while (dRead.Read())
                 {
-                    g = GetGradeFromRow(dRead);
+                    g = dl.GetGradeFromRow(dRead);
                     break; // just the first! 
                 }
                 dRead.Dispose();
@@ -1406,55 +1231,10 @@ namespace SchoolGrades.DbClasses
             return t;
         }
 
-        internal List<Couple> GetGradesOldestInClass(Class Class,
-            GradeType GradeType, SchoolSubject SchoolSubject)
-        {
-            List<Couple> couples = new List<Couple>();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                string query;
-                query = "SELECT Classes_Students.idStudent" +
-                ",MAX(timestamp) AS InstantLastQuestion" +
-                " FROM Classes_Students LEFT JOIN Grades" +
-                " ON Classes_Students.idStudent = Grades.idStudent" +
-                " WHERE Classes_Students.idClass =" + Class.IdClass +
-                " AND Grades.idGradeType = '" + GradeType.IdGradeType + "'" +
-                " AND Grades.idSchoolSubject='" + SchoolSubject.IdSchoolSubject + "'" +
-                " AND Grades.idSchoolYear='" + Class.SchoolYear + "'" +
-                " OR Grades.idGrade IS NULL" + // takes those that haven't any grades
-                " GROUP BY Classes_Students.idStudent" +
-                " ORDER BY InstantLastQuestion DESC;"; // ???? DESC o no 
-                cmd.CommandText = query;
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Couple c = new Couple();
-                    // we give back also the nulls, as Nows
-                    DateTime now = System.DateTime.Now;
-                    c.Key = (int)dRead["IdStudent"];
-                    if (!dRead.IsDBNull(1))
-                        c.Value = SafeDb.SafeDateTime(dRead["InstantLastQuestion"]);
-                    else
-                        c.Value = now;
-                    couples.Add(c);
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return couples;
-        }
-
         internal object GetWeightedAveragesOfStudent(Student currentStudent, string stringKey1, string stringKey2, DateTime value1, DateTime value2)
         {
             //throw new NotImplementedException();
             return null;
-        }
-
-        internal object GetGradesWeightedAveragesOfStudent(Student currentStudent, string stringKey1, string stringKey2, DateTime value1, DateTime value2)
-        {
-            throw new NotImplementedException();
         }
 
         internal DataTable GetWeightedAveragesOfClass(Class Class,
@@ -1479,44 +1259,6 @@ namespace SchoolGrades.DbClasses
                 " GROUP BY Students.idStudent" +
                 " ORDER BY GradesFraction ASC, lastName, firstName, Students.idStudent;";
                 // !!!! TODO change the query to include at first rows also those students that have no grades !!!! 
-                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                DataSet DSet = new DataSet("ClosedMicroGrades");
-
-                DAdapt.Fill(DSet);
-                t = DSet.Tables[0];
-
-                DAdapt.Dispose();
-                DSet.Dispose();
-            }
-            return t;
-        }
-
-        internal DataTable GetGradesWeightedAveragesOfClass(Class Class, string IdGradeType,
-            string IdSchoolSubject, DateTime DateFrom, DateTime DateTo)
-        {
-            DataTable t;
-            using (DbConnection conn = dl.Connect())
-            {
-                string query = "SELECT Grades.idGrade, Students.idStudent,lastName,firstName," +
-                " SUM(Grades.value * Grades.weight)/SUM(Grades.weight) AS 'Weighted average'" +
-                // weighted RMS (Root Mean Square) as defined here: 
-                // https://stackoverflow.com/questions/10947180/weighted-standard-deviation-in-sql-server-without-aggregation-error
-                ",SQRT(SUM(Grades.weight * SQUARE(Grades.value)) / SUM(Grades.weight) - SQUARE(SUM(Grades.weight  * Grades.value) / SUM(Grades.weight))) AS 'Weighted RMS'" +
-                ",COUNT() 'Grades Count'" +
-                " FROM Grades" +
-                " JOIN Students" +
-                " ON Students.idStudent=Grades.idStudent" +
-                " JOIN Classes_Students" +
-                " ON Classes_Students.idStudent=Students.idStudent" +
-                " WHERE Classes_Students.idClass =" + Class.IdClass +
-                " AND Grades.idSchoolYear='" + Class.SchoolYear + "'" +
-                " AND Grades.idGradeType = '" + IdGradeType + "'" +
-                " AND Grades.idSchoolSubject = '" + IdSchoolSubject + "'" +
-                " AND Grades.Value > 0" +
-                " AND Grades.Timestamp BETWEEN " + SqlVal.SqlDate(DateFrom) + " AND " + SqlVal.SqlDate(DateTo) +
-                " GROUP BY Students.idStudent" +
-                " ORDER BY lastName, firstName, Students.idStudent;";
-
                 DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
                 DataSet DSet = new DataSet("ClosedMicroGrades");
 
@@ -1574,130 +1316,6 @@ namespace SchoolGrades.DbClasses
                 DSet.Dispose();
                 return DSet.Tables[0];
             }
-        }
-         
-        /// <summary>
-        /// Gets all the grades of a students of a specified IdGradeType that are the sons 
-        /// of another grade which has value NOT null AND NOT equal to zero
-        /// </summary>
-        /// <param name="IdStudent"></param>
-        /// <param name="IdSchoolYear"></param>
-        /// <param name="IdGradeType"></param>
-        /// <param name="IdSchoolSubject"></param>
-        /// <returns></returns>
-        internal DataTable GetMacroGradesOfStudentClosed(int? IdStudent, string IdSchoolYear,
-            string IdGradeType, string IdSchoolSubject)
-        {
-            DataTable t;
-            using (DbConnection conn = dl.Connect())
-            {
-                string query = "SELECT idGrade, idStudent, value, idSchoolSubject," +
-                    "weight, cncFactor, idSchoolYear, datetime(timestamp), idGradeType, " +
-                    "idGradeParent,idQuestion" +
-                " FROM Grades" +
-                " WHERE Grades.idStudent =" + IdStudent +
-                " AND Grades.idSchoolYear='" + IdSchoolYear + "'" +
-                " AND Grades.idGradeType = '" + IdGradeType + "'" +
-                " AND Grades.idSchoolSubject = '" + IdSchoolSubject + "'" +
-                " AND Grades.Value > 0" +
-                " ORDER BY datetime(Grades.timestamp) Desc;";
-
-                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                DataSet DSet = new DataSet("ClosedMicroGrades");
-
-                DAdapt.Fill(DSet);
-                t = DSet.Tables[0];
-
-                DAdapt.Dispose();
-                DSet.Dispose();
-            }
-            return t;
-        }
-
-        internal int CreateMacroGrade(ref Grade Grade, Student Student, string IdMicroGradeType)
-        {
-            int key = NextKey("Grades", "idGrade");
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                // find the type of the macro grade of this micrograde
-                cmd.CommandText = "SELECT IdGradeTypeParent" +
-                    " FROM GradeTypes WHERE idGradetype='" + IdMicroGradeType + "';";
-                string IdMacroGrade = (string)cmd.ExecuteScalar();
-
-                // Get the Default Weight of that Grade Type
-                cmd.CommandText = "SELECT defaultWeight " +
-                    "FROM GradeTypes " +
-                    "WHERE idGradeType='" + IdMicroGradeType + "'; ";
-                double weight = (double)cmd.ExecuteScalar();
-
-                // add macrograde
-                cmd.CommandText = "INSERT INTO Grades " +
-                "(idGrade,idStudent,idGradeType,weight,cncFactor,idSchoolYear,timestamp,idSchoolSubject) " +
-                "Values (" + key + "," + Student.IdStudent +
-                ",'" + IdMacroGrade + "'" +
-                "," + SqlVal.SqlDouble(weight) + "" +
-                ",0" +
-                ",'" + Grade.IdSchoolYear + "','" +
-                System.DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss").Replace('.', ':') + "'" +
-                ",'" + Grade.IdSchoolSubject +
-                "');";
-                cmd.ExecuteNonQuery();
-
-                cmd.Dispose();
-            }
-            return key;
-        }
-
-        internal int? SaveMicroGrade(Grade Grade)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                // create a new micro assessment in grades table
-                if (Grade == null || Grade.IdGrade == null|| Grade.IdGrade == 0)
-                {
-                    Grade.IdGrade = NextKey("Grades", "idGrade");
-                    cmd.CommandText = "INSERT INTO Grades " +
-                    "(idGrade, idGradeType, idGradeParent, idStudent, value, weight, " +
-                    "cncFactor,idSchoolYear, timestamp, idQuestion,idSchoolSubject) " +
-                    "Values (" + Grade.IdGrade +
-                    ",'" + SqlVal.SqlString(Grade.IdGradeType) + "'" +
-                    "," +  SqlVal.SqlInt(Grade.IdGradeParent.ToString()) + "" +
-                    "," + Grade.IdStudent + "" +
-                    "," + SqlVal.SqlDouble(Grade.Value) + "" +
-                    "," + SqlVal.SqlDouble(Grade.Weight) + "" +
-                    "," + SqlVal.SqlDouble(Grade.CncFactor) + "" +
-                    ",'" + SqlVal.SqlString(Grade.IdSchoolYear) + "'" +
-                    ",'" + System.DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss").Replace('.', ':') + "'" +
-                    "," +  SqlVal.SqlInt(Grade.IdQuestion.ToString()) + "" +
-                    ",'" + Grade.IdSchoolSubject + "'" +
-                    ");";
-                }
-                else
-                {
-                    cmd.CommandText = "UPDATE Grades " +
-                    "SET" +
-                    " idGrade=" +  SqlVal.SqlInt(Grade.IdGrade.ToString()) + "" +
-                    ",idGradeType='" + SqlVal.SqlString(Grade.IdGradeType) + "'" +
-                    ",idGradeParent=" +  SqlVal.SqlInt(Grade.IdGradeParent.ToString()) + "" +
-                    ",idStudent=" +  SqlVal.SqlInt(Grade.IdStudent.ToString()) + "" +
-                    ",idSchoolYear='" + SqlVal.SqlString(Grade.IdSchoolYear) + "'" +
-                    ",timestamp='" + SqlVal.SqlString(((DateTime)Grade.Timestamp).ToString("yyyy-MM-dd HH:mm:ss")) + "'" +
-                    ",idQuestion='" +  SqlVal.SqlInt(Grade.IdQuestion.ToString()) + "'" +
-                    ",idSchoolSubject='" + SqlVal.SqlString(Grade.IdSchoolSubject) + "'" +
-                    ",value=" + SqlVal.SqlDouble(Grade.Value) + "" +
-                    ",weight=" + SqlVal.SqlDouble(Grade.Weight) + "" +
-                    ",cncFactor=" + SqlVal.SqlDouble(Grade.CncFactor) + "" +
-                    " WHERE idGrade=" + SqlVal.SqlInt(Grade.IdGrade.ToString()) +
-                    ";";
-                }
-
-                cmd.ExecuteNonQuery();
-
-                cmd.Dispose();
-            }
-            return Grade.IdGrade;
         }
 
         internal List<Student> GetStudentsAndSumOfWeights(Class Class,
@@ -1792,7 +1410,7 @@ namespace SchoolGrades.DbClasses
                     ";";
                 cmd.ExecuteNonQuery();
                 // crea un nuovo voto copiato dalla riga passata
-                int codiceVoto = NextKey("Grades", "idGrade");
+                int codiceVoto = dl.NextKey("Grades", "idGrade");
 
                 // aggiunge il voto copiato dalla riga passata
                 cmd.CommandText = "INSERT INTO Grades " +
@@ -1825,7 +1443,7 @@ namespace SchoolGrades.DbClasses
         {
             // trova una chiave da assegnare alla nuova domanda
             Question q = new Question();
-            q.IdQuestion = NextKey("Questions", "idQuestion");
+            q.IdQuestion = dl.NextKey("Questions", "idQuestion");
             using (DbConnection conn = dl.Connect())
             {
                 string imageSenzaHome = q.Image;
@@ -2269,7 +1887,7 @@ namespace SchoolGrades.DbClasses
         internal int? CreateNewTag(Tag CurrentTag)
         {
             // trova una chiave da assegnare alla nuova domanda
-            CurrentTag.IdTag = NextKey("Tags", "IdTag");
+            CurrentTag.IdTag = dl.NextKey("Tags", "IdTag");
             using (DbConnection conn = dl.Connect())
             {
                 DbCommand cmd = conn.CreateCommand();
@@ -2362,7 +1980,7 @@ namespace SchoolGrades.DbClasses
         internal int CreateAnswer(Answer currentAnswer)
         {
             // trova una chiave da assegnare alla nuova domanda
-            int codice = NextKey("Answers", "idAnswer");
+            int codice = dl.NextKey("Answers", "idAnswer");
             using (DbConnection conn = dl.Connect())
             {
                 DbCommand cmd = conn.CreateCommand();
@@ -3547,7 +3165,7 @@ namespace SchoolGrades.DbClasses
             int nextId;
             using (DbConnection conn = dl.Connect())
             {
-                nextId = NextKey("Topics", "idTopic");
+                nextId = dl.NextKey("Topics", "idTopic");
 
                 DbCommand cmd = conn.CreateCommand();
                 // aggiunge la foto alle foto (cartella relativa, cui verrà aggiunta la path delle foto)
@@ -3995,7 +3613,7 @@ namespace SchoolGrades.DbClasses
             int key;
             using (DbConnection conn = dl.Connect())
             {
-                key = NextKey("Lessons", "idLesson");
+                key = dl.NextKey("Lessons", "idLesson");
                 Lesson.IdLesson = key;
                 // add new record to Lessons table
                 DbCommand cmd = conn.CreateCommand();
@@ -4307,7 +3925,7 @@ namespace SchoolGrades.DbClasses
                 string query;
                 if (Image.IdImage == 0)
                 {
-                    Image.IdImage = NextKey("Images", "IdImage");
+                    Image.IdImage = dl.NextKey("Images", "IdImage");
                     query = "INSERT INTO Images" +
                     " (idImage, imagePath, caption)" +
                     " Values (" + Image.IdImage + ",'" +
@@ -4536,7 +4154,7 @@ namespace SchoolGrades.DbClasses
                 DbCommand cmd = conn.CreateCommand();
                 if (TestToSave.IdTest == 0 || TestToSave.IdTest == null)
                 {   // create new record
-                    int nextId = NextKey("Tests", "idTest");
+                    int nextId = dl.NextKey("Tests", "idTest");
 
                     cmd.CommandText = "INSERT INTO Tests " +
                     "(idTest,name,desc,IdSchoolSubject,IdTestType,IdTopic" +
@@ -4628,7 +4246,7 @@ namespace SchoolGrades.DbClasses
                 }
                 else
                 {   // create answer
-                    int nextId = NextKey("StudentsAnswers", "IdStudentsAnswer");
+                    int nextId = dl.NextKey("StudentsAnswers", "IdStudentsAnswer");
 
                     cmd.CommandText = "INSERT INTO StudentsAnswers " +
                     "(idStudentsAnswer,idStudent,idAnswer,studentsBoolAnswer," +
@@ -4841,7 +4459,7 @@ namespace SchoolGrades.DbClasses
                     Annotation.InstantTaken = DateTime.Now;
                     Annotation.IsActive = true; 
                     // create answer on database
-                    int? nextId = NextKey("StudentsAnnotations", "IdAnnotation");
+                    int? nextId = dl.NextKey("StudentsAnnotations", "IdAnnotation");
                     Annotation.IdAnnotation = nextId;
 
                     query = "INSERT INTO StudentsAnnotations " +
